Move Timeline to footer instead of header
